able to remove gene sets and reflect
